feat: add multi-page demo routing with context
Introduce hash-based router, DemoProvider, and split examples into dedicated pages.
